+++ lib/Screens/Dashboard/Jobs/Components/handyman_drawer.dart
@@ -190,40 +190,34 @@ class HandymanDrawer extends StatelessWidget {
                     ),
                   ),
                   SizedBox(height: 20 * screenHeight),
-
                   GestureDetector(
                     onTap: (){
-
-                      Alert(
+                      showDialog(
                         context: context,
-                        type: AlertType.warning,
-                        title: "ACCOUNT DELETION",
-                        style: AlertStyle(
-                            titleStyle: TextStyle(fontWeight: FontWeight.w800),
-                            descStyle:
-                            TextStyle(fontWeight: FontWeight.w400, fontSize: 18)),
-                        desc: "Do you want to delete your account",
-                        buttons: [
+                        builder: (BuildContext context) {
+                          return AlertDialog(
+                            title: Text("Confirm Deletion"),
+                            content: Text("Are you sure you want to delete your account?  This action cannot be undone."),
+                            actions: <Widget>[
+                              TextButton(
+                                child: Text("Cancel"),
+                                onPressed: () {
+                                  Navigator.of(context).pop(); // Close the dialog
+                                },
+                              ),
+                              TextButton(
+                                child: Text("Delete"),
+                                onPressed: () {
+                                  // Call the account deletion service
+                                  AccountDeletionService().deleteAccount(context);
+                                  Navigator.of(context).pop(); // Close the dialog
+                                },
+                              ),
+                            ],
+                          );
+                        },
+                      );
 
-                          DialogButton(
-                            onPressed: () async {
-                              // Call the deleteAccount method here
-                              await AccountDeletionService().deleteAccount(context);
-
-                            },
-                            color: Color(0xFF0D47A1),
-                            border: Border.all(color: Color(0xffe5f3ff)),
-                            child: Text(
-                              "Yes",
-                              style: TextStyle(
-                                  color: Colors.white,
-                                  fontSize: 16,
-                                  fontFamily: 'DM-Sans',
-                                  fontWeight: FontWeight.bold),
-                            ),
-                          )
-                        ],
-                      ).show();
 
                     },
                     child: Row(
@@ -243,7 +237,9 @@ class HandymanDrawer extends StatelessWidget {
                       ],
                     ),
                   ),
-                  SizedBox(height: 20 * screenHeight),
+                  SizedBox(height: 27 * screenHeight),
+
+
 
                   GestureDetector(
                     onTap: signOut,
